Added changes to moves' damage calculations.
