+++ idk/src/main/java/com/treydev/idk/SpecialAttack.java
@@ -1,10 +1,6 @@
-package com.treydev.idk.location;
+package com.treydev.idk;
 
 import java.util.ArrayList;
-
-import com.treydev.idk.Critter;
-import com.treydev.idk.Element;
-import com.treydev.idk.Move;
 
 public class SpecialAttack extends Move {
     //TODO: Create separate class that this and physical attack extend!
@@ -31,8 +27,8 @@ public class SpecialAttack extends Move {
     private void dealDamage(ArrayList<String> outputs, Critter target, Critter user) {
         outputs.add("You hit!");
         //TODO: (optional) move actual damage calculation to yet another equation so the equation can be overridden.
-        int attack = user.species.getBaseSpecial();//TODO: replace base attack with actual attack!
-        int defense = target.species.getBaseSDefence();
+        int attack = user.getSpecial();//TODO: replace base attack with actual attack!
+        int defense = target.getSpecDef();
         int level = 50;//TODO: REPLACE WITH ACTUAL LEVEL!
         int damage = Power;
         damage *= attack;
